Defending against reentrancy
Especially in the case of augmenting console, it is possible that a subscriber function could inadvertently trigger a publish, which could lead to an infinite loop. Since subscribers are intended to be transparent and not impact the flow of the application significantly, we ignore any publish events triggered directly by a subscriber. It is still possible to get an infinite cycle if a subscriber does something asynchronously.