Adding parameter to subscriber filters
This parameter allows the subscriber filter to know whether this is a speculative publish or not.
Prior to this change, a filter trying to implement sampling might look like the following

    () => { return Math.random() < 0.1; }

However if a publisher attempted to avoid additional work if there are no interested subscribers,
this would only report %1 of events rather than %10.

The filter can fix this issue by caching the previous value, but in the presence of multiple subscribers
the filter would not know whether it is being asked about a new event, or the same event which another
subscriber was interested in.

With this change, the following works:

    var lastValue = true;
    var filter = (publishing) => {
        if (publishing) {
           return lastValue;
        }
        lastValue = Math.random() < 0.1;
        return lastValue;
    }

The filter will now return the same value during the publish call that it did in the previous shouldPublish call

Note that if the publisher does work between calling shouldPublish and publish which invokes the same subscriber,
then all bets are off.